[Requirement #9536] - Ensure Build Script repository is managed properly
Updated the manifest generation so that it allows generating a manifest for
a directory that has spaces in it.  Also, turned off checksum checks on the
Logo & splash screen in the Vet edition.  Finally, updated the vetimageserver
solution file to include several of the manifest input files that are used and
were missing.