[Bug #9568] - Association aborted due to timeout when image takes too long to transfer
Updates to the ClearCanvas.Dicom.Samples project as part of the ticket.  Removed
the Logger, and adapted read the log4net logs and show them in the log window.
Also removed the StorageScu & VerificationScu classes and used the ones in the
dicom library, due to log4net logging now going right to the log window.  Also
made some ReSharper updates along the way.