[Defect #9584] - Spacing between partition and search field changes after clicking search
- Fixed. The issue was that although the fields are hidden, the parents (table cells) are still visible. This caused the problem in IE.

- Also refactor the code a little bit to move the code that turn on the vet fields into the plugin. Although this doesn't buy us much as it stands
but it's better than leaving the code in the open source. In addition, (if we decide to do so) this will allow us to turn on the fields all the time
instead of relying on the presence of the Vet authority tokens (and expect the admin remember that everytime they configure the system).