[Defect #9683] - Data Access authority token should be sorted [Defect #9686] - All authority groups with access to a study should be listed in Study Details Data Access tab
Both are fixed.  The data access tab now shows two lists: (data access) groups given access to the study through the data access assignment, and all other groups that can access the study via the token.
If the first list is too long (>10), a link is shown on top for user to jump to the 2nd list if it's visible. The lists are both sorted.

Note: Some existing classes and methods have been renamed for clarity.