[Bug #8506] - Image level query matches on Specific Character Set
Note, the patch submitted for this defect didn't work due to some changes in the
code since 2.0.  Note that for some reason, a change was made in the 2.2
code to remove a change on Query Retrieve Level for instance queries.  In any
case, added in that we ignore the specific character set tag in C-FIND-RQ at
the image level, and then fixed a small bug with encoding the c-find-rsp
at the image level from the source image.