[Requirement #9782] - The Workstation will continue to support basic DICOM functionality as a PACS viewer
Initial implementation of the WorkItemService for handling calls from the
Desktop.  Note that the shells for these classes are here, and there's not
a full implementation.  The Subscribe/Unsubscribe for updates
mechanism uses the SubscriptionManager, but this isn't really a great fit right
now..  This needs some work, but its a basic implementation.