[Requirement #9782] - The Workstation will continue to support basic DICOM functionality as a PACS viewer
Addition of a WorkItemProcessor class for processing the Workitem queue.  Added
some simple interfaces and extension points for creating processors for the
various types of workitems.  Also, added a WorkItemStatusProxy class for
updating the status of WorkItems.  This class can be where we wire in the
progress updates that can be sent to the GUI.