[Requirement #9806] - Imported Studies shall be placed in the WorkQueue for processing
First run through with updates to get importing working:
 * Hacked some code together to communicate with the server to insert a
WorkItem request.  This works, but we need to rework the code.  Its used
right now by the DicomFileImportTool.  Ie, the WorkItemMonitor needs to be
reworked, plus we need to think if the callback mechanism for WOrkItem updates
should be pulled out of IWorkItemService and be its own service.
 * There's an issue with the ImportFilesRequest serialization.  the list of
files to import is not being deserialized properly, and the WorkItem processor
is failing to intialized because it can't deserialize the list.