[Requirement #9806] - Imported Studies shall be placed in the WorkQueue for processing
Interim Commit.

Modifications to the SopInstanceImporter to be able to pass in the
bad file and import file behavior.  Included adding a new CopyFileCommand
in CC.Dicom.utilities.Command for copying a file.

Also made some small updates related to the handling status updates in the
work items.

Finally, pluged in the processing of status updates into the WorkItemActivityCallbacks.
This required adding to IWorkItemActivityCallback a known type provider to
deal with serialization properly.

Note, there's still some issues with the Progress object not having its status
relayed properly to the client monitoring the status.  Plus, more general
unit testing to be done.