[Requirement #9782] - The Workstation will continue to support basic DICOM functionality as a PACS viewer
Updates to make Progress updates work.  I was mistakenly deserializing the progress
item each time it was updated, erasing prior updates. In process of cleaning
this up, reworked the IWorkItemProcess interface, and made updates to the
BaseItemProcessor to automatically handle the progress and request objects
and to intialize them appropriately.