[Requirement #9782] - The Workstation will continue to support basic DICOM functionality as a PACS viewer
Introduced WorkItemStudyRequest has a parent class for all Workitems that work
on a study.  This request then has the Patient & Study objects and will
make it easier for the GUI to grab the patient & study.  Also, changed
WorkItemData's Patient & Study properties to grab from the request, if it
exists.

Also, found a minor publishing issue where we were not publishing when we
insert import study items.

Finally, updated the SopInstanceImporter's context objects, so that a user
can listen for events when a new WorkItem is inserted to help with publishing.