[Requirement #9782] - The Workstation will continue to support basic DICOM functionality as a PACS viewer
Added Publish method to IWorkItemService and added Refresh to the
IWorkItemActivityMonitorService service.  This includes implementation.

Made updates to include publishing progress when inserting Import Study
requests & receiving studies over the network.

Made some updates to how the StoreScpExtension keeps a context for importing.

Found some problems with the InsertWorkItemCommand that became apparent when
saving the context to be reused over a DICOM association.  Also cleaned up
an issue with the FileUtils.Backup() routine not trying new filenames on a
name collision.