[Requirement #9806] - Imported Studies shall be placed in the WorkQueue for processing
Cleaned up a progress issue with import after the prior commit.  Also,
updated the SopInstanceProcessor so it can be passed files in batch to be
processed.  Defaulted the batch to 10 SOPs right now.  its still transaction
based across the 10 SOPs.

Note, there's some logging cleanup and some other general documentat/comments
to bge done in the code yet, but it looks like it works and its faster,
with the exception of duplicate imports, which i have some ideas on cleaning up.