[Requirement #9806] - Imported Studies shall be placed in the WorkQueue for processing
Additional code cleanup.  Added comments in some areas, did some general
renaming and refactoring, and cleaned up the code in general.

Also, made the batch size when processing configurable, and made a change to
improve performance when dealing with duplicates.  We now overwrite the
instance via a copy when dealing with a duplicate instead of saving the
DICOM file in memory.