[Requirement #9806] - Imported Studies shall be placed in the WorkQueue for processing
Cleaned up a number of bugs in the previously committed Reindex code.  The biggest
was a temp file issue with leaving temporary files when writing xml files
around in:

C:\Users\<user>\AppData\Local\Temp

Note that other dcm files may be left around in that folder also.

A basic reindex is operational, but it has to be tested a lot more, plus I have
to investigate speeding it up, and potentially reducing the logging.  Also,
introduced some checking in the other WorkItem processors to see if a reindex
is scheduled to cause delays.