[Requirement #9811] - Re-index operations will be inserted into the WorkQueue for processing
Updated the insert of WorkItems, so that if an Reindex is requested, and one is
already in progress or operational, we'll just return the existing reindex
request, instead of inserting a new one.

Also, made some updates related to the configuration options for the service
and how we set the expiration and scheduled times.  This still needs a bit
of work.