[Requirement #9806] - Imported Studies shall be placed in the WorkQueue for processing [Requirement #9811] - Re-index operations will be inserted into the WorkQueue for processing
General cleanup of reindex and reprocessing studies.  Change the reindex on
reprocess to just rebuild the study xml file, and not to process every file
and flush the study xml to disk.  Introduced a new ReprocessStudy class
that is used by the ReindexProcessor to make this work.  Also made some general
cleanup in testing.