[Requirement #9782] - The Workstation will continue to support basic DICOM functionality as a PACS viewer
Further attempts to elimnate the locking that's occuring with the database.
Tried making it so when we batch together the processing, that it does all the
database updates together at the end of the process, but this didn't seem to make
a difference.  Simplified down the batch so we only committed a single update
at a time, and the timeouts still occurred.  Settled on batching 25 instead of 50
studies at a time.  Still need to look into more options to resolve this.