[Requirement #9798] - Activity Monitor
 * replaced the call to Query with a call to Refresh, so that the component now relies entirely on the refresh mechanism.

Note: This seems to have degraded the UI responsiveness to filter changes, presumably because it now has to wait for work item updates to arrive, and that seems to take some time particularly when the shredhost is busy doing other processing.  Not sure this is really the way to go.