[Requirement #9817] - Auto-route rules will be applied once a Work Item has completed processing
Performed some more refactoring of the IS Rules Engine to share more code with the IV and created an initial auto-route:
 * Moved ServerActionItemBase to CC.Dicom.Utilities.Rules and made generic & renamed ActionItemBase
 * Moved ActionOperatorCompilerBase to CC.Dicom.Utilities.Rules
 * Moved DicomAgeGreaterThanSpecification to CC.Dicom.Utilities.Rules
 * Moved DicomAgeLessThanSpecification to CC.Dicom.Utilities.Rules
 * Moved DicomExpression to CC.Dicom.Utilities.Rules.

Also, created AutoRouteActionItem & AutoRouteActionOperator shells based on the
IS ones.  These need cleanup.

Also, cleaned up some Webstation related compilation errors in the IS related
to the refactoring in the IV how AE Titles are dealt with.