[Requirement #9817] - Auto-route rules will be applied once a Work Item has completed processing
Did a bit more work on the rules engine.  Added a shell Load() routine to the
ViewerRulesEngine that would show how to load up and populate the rules engine
and also compile each rule.

Also, introduced the transferSyntaxUid parameter to the auto-route action.  note
that we probably could change the "device" parameter to be deviceOid to identify
the configured device.

Also, did a bit of work with inserting the Auto-route into the database.