[Requirement #9807] - DICOM Send from the UI shall be placed in the WOrkQueue for processing
General updates to support DICOM Sends through the WorkQueue.  Note that a
bunch of code cleanup still needs to be done. Added a new DicomSendItemProcessor
and factory to handle the server side processing.  Introduced a
ImageViewerStorageScu that wraps the existing StorageScu and adds in some
additional support for loading the list of files to be sent.  Also did a
preliminary implementation of the dicomSendItemProcessor.  Needs a bunch of
testing and cleanup, but wanted to get it commited for now.