[Requirement #9807] - DICOM Send from the UI shall be placed in the WorkQueue for processing
Cleaned up the exception handling in the WorkItemClient classes, they now throw
exceptions instead of returning true/false in some cases.  This seems to integrate
a bit better with the uses in the GUI.  Alos, this commit includes support for
doing a DICOM send from the GUI.

Also, added coded for publishing files, specifically for publishing when streaming
from a server.