[Requirement #9831] - Failed items in the WorkQueue can be retried
Realized the deletion of workitems was not being handled cleanly after
introducing the delete from the GUI.  Had to introduce a new
WorkItemStatusEnum, DeleteInProgress to address the issues.  We now set to
this status while the WorkItemProcess performs the actual delete, so that the
WorkItem is "locked" while its being deleted.  The WorkitemProcessor must
do the delete so that if we have spare files due to failure on the filesystem,
we actualy remove them.