[Requirement #9817] - Auto-route rules will be applied once a Work Item has completed processing [Requirement #9952] - The user can manually re-apply a study rule
Cleaned up an issue where Auto-routes were not being published to the Activity Monitor,
this required making the WorkItemActivityPublisher a public class.  note there's still an
issue that we insert a WorkItem in CC.IV.Dicom.Core, and we'd have a circular reference
if we tried to publish from there.

Also as per tech review, cleaned up the labels for Send Series/Study Delete Series/Study
to include the A# & study date.