[Requirement #9964] - Manual retrieves from the UI will be placed in the Work Queue for processing [Requirement #9947] - The user can cancel import work items before they start
Updated to support retrieves.  Implemented the WorkItem processor for retrieves,
and the WorkItemClient.  This includes Series & Study level retrieves.  Also,
found a small issue with the ServerDirectory that it didn't work properly when
a match was not found when searching by name,and also cleaned up an
issue with DicomServiceNode where it wasn't returning the our existing server
node with the proper name.

Finally, fixed an issue where we couldn't cancel WorkItems until they ran at
least once.