[Requirement #9782] - The workstation will continue to support basic DICOM functionality as a PACS viewer
More code cleanup.  Reworked how Expiration times work on WorkItems to be more
like the ImageSever.  The time is now the # of seconds on an import since the
last WorkitemUid inserted.  Also, there were a few cases where the expiration
time was hard coded, and not using the configuration option.  Updated this also.
Finally, added a bit better logging of when expiration times are for WorkItems,
note that in the ifnal rev this will probalby be turned off.