[Requirement #9817] - Auto-route rules will be applied once a WOrk Item has completed processing [Requirement #9805] - SOPs will be compressed on-the-fly when being auto-routed
Implementation of feedback from sprint review.

Implemented changes in the CanStart() method for Dicom Sends, so that we don't start
a send if there's an existing WorkItem for the study in progress, or if a
ProcessStudy entry is Pending, Idle, or In Progress.

Also, updated the labeling for importing files as per feedback to be "Process
files for import from {0} location(s)"