[Requirement #9817] - Auto-route rules will be applied once a WOrk Item has completed processing
Changed the scheduling of auto-routes, so we preserve the Start/End times
in the WorkItem request.  We also now check to make sure when we execute an
auto-route request that we're in the time window, and reschedule ourselves if
we're not.  Finally, added some unit tests for calculating the scheduled time.