[Requirement #9813] - Created SOPs will be published via the WorkQueue
Reworked publishing so that it no longer has the middle step of having an
"Import Files" Workitem.  Instead, it goes directly to an Import Study item.
This also required moving the DicomFilePublisher class into
the IV.Dicom.Core assembly, which is probably more where it belongs anyways.
This assembly now must be included in the desktop build.  Finally, fixed a small
thing from the text string when importing files.