[Requirement #9782] - The workstation will continue to support basic DICOM functionality as a PACS viewer [Requirement #9817] - Auto-route rules will be applied once a WOrk Item has completed processing
As per sprint review, added tool tip over the status field with the extended failure
message.  IE, when a failure occurs, just however over the status and it will be
shown.

Also, several other cleanups: Changed when we insert a workitem, the expiration time
is based on the first scheduled time, instead of from "now".  Also, fixed an issue
so that Stat auto-routes can run right away, instead of in the time window.