[Requirement #9832] - The user can delete Pending, COmplete, Cancelled, or Failed work items
As per a test report, cleaned up how we handle updating of WorkItems.  the
code attempts to load the WorkItem to be updated, then update it.  it was not
checking if the WorkItem to update was not found, and a null exception was
happening.  Added a check if the WorkItem could not be found.