[Requirement #9794] - WOrkstation local services will be refactored to use a WorkQueue
Added back in Publish method on IWorkItemActivityMonitorService.  This was
previously on IWorkItemService, but it makes sense to have it on the same
interface that you subscribe/unsubscribe to updates.  Note that this was
required because the publishing was occurring from different app domains in the
ShredHostService, and the publishing was not working properly and required the
service.

As part of this, I had to change some updates to Progress for importing studies
to save each Progress update to the database.  This in turn, helped make updates
to progress when receiving over the network to be more up to date.