[Requirement #9972] - The Workstation will target .NET Framework 4
Updated nunit to version 2.6.0.12051.  Note that I had to download the source to
this version, update the solution file to VS2010, reset the target to
.NET 4.0, then compile a release build.

Note also did a test run with running some unit tests, and cleaned up a minor
issue with a failure in the DICOM library unit tests.