[Requirement #9972] - The Workstation will target .NET Framework 4
Updated the ImageServer solution files to use .NET 4.0.  Note that there were a
couple changes required to fix the build related to IV changes for Marmot.  Also,
this includes updates to Kakadu 7.0.

There are still references to the old 1.0 version of the AJAX extensions,
including the Validators.dll.  This should be updated to the latest version in
.NET.