[Requirement #9972] - The Workstation will target .NET Framework 4
Reworked the Jin DicomTestTool to use the .NET thread pool to do compression,
also reworked how it reports errors and problems so it does them in a batch after
everything has been processed.  Finally, added functionaly to scan a
directory and run the comparison tests on all dicom files found in the
directory, recursively.