[Requirement #9807] - DICOM Send from the UI shall be placed in the WorkQueue for processing
Found issue with DICOM file where the meta info (group 0x0002) did not match the
SOP Class UID tag in the data set.  This was causing an error in the SCU where we
failed to send a file over the network.  made it so we select the presentation context
based on the SOP Class UID tag.