[Defect #9968] - Viewer can get into infinite retry loop when streaming images with LUT encoding error
The workstation did not get into infinite retry loop. It just tries to retrieve images repeatedly and only gives up after trying for about 2 seconds.
During this period, the workstation becomes less responsive. This happens every time the workstation tries to render an image or prefetch it.
The problem gets worse as number of image boxes increased and when the window is resized.

* Changed FramePixelDataRetriever so that it does not try to retrieve again if initial attempts already failed.  The same error that previously caused it to fail would be thrown immediately and rendered on the screen.