[Requirement #9809] - Manual Series Deletion from the UI shall be placed in the WorkQueue for processing
Changed the constructor to InsertOrUpdateStudyCommand to no longer pass in a
DicomMessage, since we are just using the StudyXml.  This change made it so
when we process a batch of images withina study, we only need to have
InsertOrUpdateStudyCommand to be executed once for the study, instead of having
it execute once for each image in the batch.  This should improve performance a
bit.