[Requirement #9811] - Re-index operations will be inserted into the WorkQueue for processing
Fixed issue with the Re-index not causing the studies list to be cleared.
The server side code was not forcing the event to be subscribed to.  Also, made
it so the title bar is updated after the reindex starts so the count is correct.

I know this worked, so not sure when it broke and stopped working.