[Requirement #9809] - Manual Series Deletion from the UI shall be placed in the WorkQueue for processing
The DeleteProgress was setting IsCancelable to false right away.  Changed it to
true, and now we update it to false when it starts being processed.  Also,
made some improvements such that if a failure occurs, or if the item is
rescheduled, the WorkItem goes back to being cancelable.  This should solve
the issues finally.