[Requirement #9764] - Workstation local services will be refactored to use a WorkQueue [Requirement #9808] - Manual Study Deletion from the UI shall be palced in the WorkQueue for processing
Further changes getting the concurrency rules working.   Reworked the
updating study concurrency item so it will not run if any other Workitem for
the study is currently in progress.  This makes things a bit more restrictive.
The side effect of this is that a study doesn't start to be processed until its
been completely received over the network.  Processing happens pretty fast, but
its not visible right away because of this.

Also, includes database updates. Removed the InsertTime column from the WorkItem
table and replaced it with the ProcessTime, the ScheduledTime is now considered
the initial insert time/order of the entries, and the ProcessTime is when
items will be checked on by the Work Item processor.

Finally, made updates to the canceling of StudyDelete WorkItems, so a study is visible
again after the deletion is being canceled.  Also, changed so that delete
study entries that have been canceled cannot be set back to pending.  i had it
this way at first, but it ended up being confusing and difficult to get right
if there were multiple delete entries, completed delete entries and the study
was received again, etc.