[Requirement #9805] - SOPs will be compress on-the-fly when being auto-routed
The code didn't work properly when changing from one lossless syntax to another.
There were issues in that the StorageScu needed to change to an uncompressed
syntax first, and that the actual ChangeSyntax routine when working on a
DICOM message, was not setting the DicomMessage to have the new transfer syntax.