[Requirement #9808] - Manual Study Deletion from the UI shall be palced in the WorkQueue for processing
Updated how related WorkItems are dealt with when a study is deleted.  At first
they were just removed, then yesterday I made an update to only remove the
non-completed ones and plublish the updates, finally, today, making it so all
the related WorkItems just stay in the WorkQueue, and they will fail on their own
if the study has been removed.