[Requirement #9807] - DICOM Send from the UI shall be placed in the WorkQueue for processing [Requirement #9964] - Manual retrieves fromt he UI will be placed in the Work Queue for processing
Updated how status is handled for sends & retrieves.  When individual error
statuses were happening, we did not preserve the status update until the failure
occurred.  Also, added in some simple extended error handling in the dicom
library to fillin the ErrorComment tag in DICOM, and for the Move SCU &
Storage SCU to pick up the error comment tag, if its set.

Note that this will require a small update to the DICOM conformance statement
for the Workstation.