[Requirement #9832] - The user can delete complete or failed work items
Changed the activity monitor so Pending WorkItems that are "Cancelable" can be
deleted when they're pending.  Note that since DICOM Receives are not
cancelable, they cannot be deleted.

Also, updated the Study Process WorkItems to go to Idle when they're not being
processed, instead of Pending.  This makes it so that the WorkItems cannot be
deleted after they start processing.

Finally, decided to update the settings so that the expire delay default value
is now 30 seconds instead of 60 seconds.  This should speed up a bit the
completion of work items.