[Requirement #9728] - Perform design reviews
After some performance monitoring, discovered that the simple opening/closing
of the database connection was taking an extensive amount of time.  Did some
reasearch on the web, and because of how we're using mutexes and loading
opening connections to the database, it appears that the sqlce connection
is reloading the database into memory each time a connection is opened.  Added
a connection object that stays open as long as the Shredhost is open, and this
appears to speed up the processing times, etc., of the code.