[Defect #10065] - Re-index starts while Receive is in progress
Updates to make it so a Re-index waits for Pending/Idle Study Insert Workitems
to complete before it starts.  It previously would only wait for the items
to be Pending or Idle, which could cause a re-index to start when a partial
study is processed.  Had to make updates to how we check for Re-indexes when
scheduling Receive Study/Import Study workitems to make this work.

Also made some updates to how we display in the AM that a WorkItem is waiting
for another Workitem to complete to make it more clearer than it was previously.