[Defect #10077] - Retrieve Work Items should not be cancelable
Made a round of updates to first implement some changes to make cancellation of
move requests work properly, then made a change to actually disable them to
match the requirements. Note that part of the issue with keeping them
non-cancelable is defect #10087 in the ImageServer.  If we make the retrieves
cancelable, it should work even with the IS bug.