[Defect #10037] - Activity Monitor move destination message should be friendlier
The Cheetah branch addressed this by adding a translator to translate status values
into a new localizable message.  Decided to address this by updating the
status descriptions in DicomStatus and then putting some of the status values
into a new string resource that can be localized. Also had to move a DICOM print
related string resource out of the Dicom.SR to Dicom.Network.SR.