[Defect #10052] - Misleading error message on retrieve when disk space exceeded
Adjusted how status values are retrieved in the DICOM toolkit & DIcomStatuses class.
Made it so service class specific error statuses can be returned, and adjusted
the FindScu & MoveScu to grab the query retrieve specific error messages.  The
bug was that we were grabbing a Storage related error message instead of the
more specific query retrieve error status/message, and the error displayed
was misleading.

Note that we still have to adjust perhaps to display a more meaningful error.