[Defect #10082] - Deleting a failed/canceled import deletes unprocessed files, but not the empty folder [Defect #10079] - Importing files for a study that previously failed to import does nothing
Reworked the code that selects a matching WorkItem, if it exists.  There was
a bug in WorkItemBroker.GetPendingWorkItem in how it checked for pending, there
was an incomplete list of status, changed so we check explicitly for the Pending
status values.

Also, made updates in cleanup of failed workitems.  For non-duplicate objects,
we now check to see if the SOP is in the studyXml before deciding to delete it.