[Defect #10046] - ShredHost service repeatedly crashes when auto-routing certain already compressed images
Fixed a specific access violation exception seen in Jpeg codec which caused the shred host to crash. The code in JPEGCODEC::Encode() used jpeg_destroy_compress()
to clean up a memory block which was never initialized by jpeg_create_compress() because an exception was thrown before it.


Important Note:

StorageScu is also updated to catch any exception in the worker thread. In theory this should prevent the shred host from
crashing if anything happens during DICOM transfer. However, this article (http://msdn.microsoft.com/en-us/magazine/dd419661.aspx)
says state corruption exceptions (SCE) such as access violation exception is treated differently in .NET 4. StorageScu will not be able to
catch SCE and Shred host will still crash if such exception is thrown in other part of the code.

As mentioned in the article, if we want to catch and handle SCE, we need to add  legacyCorruptedState­­ExceptionsPolicy=true into the app config or mark the function
with HandleProcessCorruptedStateExceptions attribute. However, it's generally safer to let the program crash if SCE occurs.