[Defect #10069] - Auto-routes scheduled within a time window aren't clearly identifiable
Fixed a bug in the Postpone routine where it didn't identify time window work
items properly and update the scheduled time.  The Scheduled Time displayed in
the Activity Monitor will now reflect the time window where its scheduled.

Did not update the code to add a details about rescheduling, this should be
sufficient.