[Defect #10044] - Can no longer import from a mapped network drive to which the service user account does not have access
Added a mechanism to cache the user credentials when an Import Files WorkItem is
either inserted, or its status is updated.  The credentials are cached for an
hour.  So, if an entry fails, and the WorkItem is reset, it will cache the
credentials again when the WorkItem is reset.

Note also that this fix required some changes to the config files for the various
editions of the workstation (including Team Edition).  This means that technically,
the fix should be tested with each edition of the workstation, to make sure the
updates were done properly.

Finally, made a few minor updates to the team solution to match it with the folders
used for the standalone imageviewer solution so they're consistent.