[Defect #10080] - Re-index starts while study import is in progress
Found an issue where if you have a re-index in process, and there's a failed
import that was scheduled before the re-index, if you reset the re-index to
pending, it would run right away, instead of waiting for the re-index to
complete.  Changed so we also will not start if any re-index is in progress.