[Defect #10071] - Anonymized study sometimes not fully imported with no errors in the Activity Monitor
made a change to detect and throw an exception when the import fails.  Still don't have
a full solution on why the import is failing.

Upon further review, made a mistake in the prior commit for this in where I
started the impersonation of the user identity.  I was only doing it when
the file was loaded, whereas more of the code needed to impersonate.