[Requirement #10105] - The user can configure the delay time and number of retries for DICOM send operations
Added new configuration database, and introduced the SystemConfigurationSettingsStore,
which stores settings associated with it in the database.  Note that I just have
some basic retrieves out of the database, and the framework for the settings
in place, i haven't worked on the settings upgrade code.  Note that this only
works for application settings, and not user settings (although there is some
code that could handle user settings).

Also implemented a new settings class, DicomSendSettings, which in turn is
stored with the new SystemConfigurationSettingsProvider, and is now used by
the DicomSentItemProcessor WorkItem processor to get settings for retries
for dicom sends.  This still has to be glued into a new application component
in the GUI.