[Requirement #10104] - The activity monitor will maintain items in sorted order
Added RequestedTime to the WorkItem table in the database.  The value is also
now filled in and populated for the activity monitor.

Also, removed the Rules table from the dicom_store database.  This table
was no longer used, but still in the database.  Also, cleaned up the db
creation scripts a bit.