[Defect #10044] - Can no longer import from a mapped network drive to which the service user account does not have access
Moved the ExclusiveLock class from CC.Standalone.Common to
CC.Common.Utilities and made it public.  The class is now used by the
DataAccessContext for the mutex to block access to the database.  This code
is known to work on XP under a user account with limited access rights.  Also,
made the SqlCeDatabaseHelper.GetDatabaseFilePath() method public, so it could
be used to name the mutex and make it unique per "family" install.  Note that
ExclusiveLock could be used by the appender in CC.Common.log4net, but decided
against doing that for now, at least.

Also, discovered there was an issue with the regression tests for the settings
store, and specifically TestSettingsStore, where the desktop app is using this
settings store when debugging instead of the actual one.  This causes changes
to the retry to not propogate to the shredhostservice.  Added a comment, but
didn't fix this.