[Defect #10053] - While importing/receiving, user can enter multiple "Delete" work items for the same study
Introduced a couple of changes, first, the Insert routine for work items
now checks for already existing Study delete entries, and doesn't insert a new
one if an existing one is found for the same study.

Also, in the process of testing, found an issue where 2 deletes for the same
study were entered, and they never ended up executing because the blocked each
other from executing.  Made a small change in the CanStart() routine to allow the
first scheduled entry to run, and the second scheduled entry waits in this case.
Note that this would be a bit hard to test, because the former changes should
prevent this from happening any more.