[Defect #10113] - When max diskspace exceeded, some status details messages are confusing [Defect #10054] - Changes made to max disk space configuration not seen by all activity types
Updated some changes done for #10054.  The changes there caused the
"Import Study" workitems to fail, when the exceeded disk space was detected
in an "Import Files" WorkItem.  This was leading to some odd behavior with
processed images count.

Updated so that the Import Files workitem only fails with an appropriate
error message, "Disk usage exceeds maximum configured usage.", which is shown
in the status details message.  Also removed, a Status details update that
was updating the files being imported as they were being imported.  This was
overwriting failure descriptions, which was losing some key details.

Also, introduced the concept of a "Fatal Error" when processing imports.
This casues the Workitem to immediately quit after its detected, instead
of continuing to attempt to import files.  This error is now triggered
when running out of disk space.