[Requirement #9728] - Perform Design Reviews
Updates based on code review for locking access to the ObservableDictionary
on the ImportFilesContext.StudyWorkItems property.  Introduced a
StudyWorkItemsSyncLock property for locking access.  Also, discovered some
issues with publishing on import.  Some recent updates for publishing anonymous
studies messed this up a bit.  We now publish only based on the
ObservableDictionary being changed, and not directly in the ImportFilesUtility.