[Defect #10126] - Series delete is audited as an update
Added a new AuditHelper.LogDeleteSeries() method for use when deleting a series.
I took a look at the IS first, and it audits a "Delete Study" audit message
when a series is deleted. The proposed solution to have the "D" event type
set does seem like a better solution since the whole study isn't deleted, and
the "study" is being accessed, but not deleted.

Also, while testing, found a small bug where the Alert being generated when
a series was being deleted was not showing the count of how many series were
deleted (it always reported 0), becaue the series list was being refreshed and
the selected count was being reset.