[Defect #10071] - Anonymized study sometimes not fully imported with no errors in the activity monitor
Discovered according to the MSDN Mutex class docs that Global\ should be
prepended to a mutex name to make it work across terminal server sessions.
It turns out our Desktop app + ShredHostService are treated as different
sessions since they run under different users, so the ExclusiveLock mutexes
were blocking within a process, but not across process boundaries.  Also
fixed this with the log file lock, since it had the same issue.

Finally, found a small issue with exceptions not being caught when progress
updates are pushed to the database in the WorkItemStatusProxy class.