[Requirement #9728] - Perform Design Reviews
Code review found an issue with settings migration and the new SystemCOnfigurationSettingsProvider.
I had some bad assumptions about how migration would work, and thought that
if the current settings group was in the settings store, then migration should
be allowed, when it shouldn't be.  Cleaned this up, cleaned up the unit tests,
plus fixed another issue where the unit tests "test settings store" was overriding
the real settings store when debugging the app.  I now use a
UnitTestExtensionFactory to set the settings store for the tests.