[Defect #10122] - Received/imported files are not audited
Cleaned up the audit logging for DicomInstancesTransferred and DataImport audit
logs.  The DicomInstancesTransferred audit log is now in the StorageScpExtension
only.  Removed this audit log from the MoveScu.

The DataImport audit log is now done from the ProcessStudyItemProcessor.  This
happens on completion or failure of the work item.  Also, introduced into
the ImportFilesUtility a DataImport audit log, but only a failure, so we capture
when we fail to import a file.  Also, cleaned out other places (such as publishing)
where we were audit logging imported studies, and leaving it to the work item
processor.